Add int insertKeyInTree() for BPlusTree
Assumed that the target records are already existing on the disk and
we are simply inserting the Key into the index tree. However, when an
internal node has (n+1) keys, the records will be moved along the index
when shifting. This is not using Syed's shift methods.

Shifting of records and index is done using insertion sort since the
nodes are already sorted, so the time complexity will be O(n).

For insertion purposes, every node has a parent pointer, even the root
which will simply point to NULL.

Crucially, I have included the numKeys and numChildNodes for the
respective struct definition of leaf and internal nodes. This is because
array is static, so there is no way to deallocate an element from an
array unless we use vector, a dynamic array. Unfortunately, vector has
its own issues so I think we should stick with static arrays. Hence,
if numKeys = 3, we will only consider the first 3 elements in the array
even if there is a 4th element in the key.