Amend InsertRecordInTree() to make use of Stack
InsertRecordInTree() now uses a stack for traversal of the BPlusTree.
Struct definition for InternalNode and LeafNode has also been changed
to reflect numKeysInserted only, removing the number of keys inserted.

Unfortunately, I am a bit uncertain for Case 2.2 and Case 2.3 for update
of parentNode as I am unsure if I am updating the keys in the parent
node accurately.